created new directory (xbrl_processing)
This directory is to be used to store all of the Python functions for processing the xbrl data once it is in a usable csv format. Functions should be stored here before intergation into the pipeline. 

Additionally added new file within directory (unique_entries.py) which contains the unique_entries function (post-review)